+++ google2/CHANGELOG.md
@@ -16,6 +16,7 @@ Added:
 - Added `createDataset` and `deleteDataset` to `GoogleBigQueryService`
 - Added new constructor to `GoogleBigQueryService` that accepts path to credentials JSON
 - Added `deletePv` to `KubernetesService`
+- Added `namespaceExists` to `KubernetesService`
 
 Dependency Updates:
 ```
@@ -26,7 +27,7 @@ Update http4s-blaze-client, http4s-circe, ... from 0.21.16 to 0.21.18 (#499) (2 
 Update sbt from 1.4.6 to 1.4.7 (#500) (2 minutes ago) <Scala Steward>
 ```
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.19-3a6194b"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.19-TRAVIS-REPLACE-ME"`
 
 ## 0.18
 Added: